requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
openai==1.50.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
pytrends==4.9.2
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.12
